requirements.txt
Flask==2.0.1
pandas==1.3.3
numpy==1.21.2
langchain==0.0.1
unstructured==0.4.0
chromadb==0.3.0
pillow==8.4.0
pydantic==1.8.2
lxml==4.6.3
matplotlib==3.4.3
tqdm==4.62.3